feat: implemented file uploads for coffee/roaster images... in theory

+++ internal/coffee/controllers.go
@@ -2,9 +2,16 @@ package coffee
 
 import (
 	"errors"
+	"fmt"
 	"net/http"
 
+	"github.com/indeedhat/barista/internal/auth"
 	"github.com/indeedhat/barista/internal/server"
+)
+
+const (
+	CoffeeImagePath  = "upload/coffee/"
+	RoasterImagePath = "upload/roaster/"
 )
 
 type Controller struct {
@@ -28,6 +35,8 @@ type createCoffeeRequest struct {
 }
 
 func (c Controller) CreateCoffee(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
 	var req createCoffeeRequest
 	if err := server.UnmarshalBody(r, &req); err != nil {
 		server.WriteResponse(rw, http.StatusBadRequest, nil)
@@ -59,6 +68,7 @@ func (c Controller) CreateCoffee(rw http.ResponseWriter, r *http.Request) {
 		Rating:    req.Rating,
 		Roaster:   *roaster,
 		Flaviours: flavours,
+		User:      *user,
 	}
 
 	if err := c.repo.SaveCoffee(&coffee); err != nil {
@@ -67,6 +77,39 @@ func (c Controller) CreateCoffee(rw http.ResponseWriter, r *http.Request) {
 	}
 
 	server.WriteResponse(rw, http.StatusCreated, createSuccessResponse{coffee.ID})
+}
+
+func (c Controller) UpdateCoffeeImage(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
+	id, err := server.PathID(r)
+	if err != nil {
+		server.WriteResponse(rw, http.StatusNotFound, nil)
+		return
+	}
+
+	coffee, err := c.repo.FindCoffee(id)
+	if err != nil {
+		server.WriteResponse(rw, http.StatusNotFound, errors.New("Coffee not found"))
+		return
+	}
+
+	// TODO: i should probably find a nice way of doing this check with middleware but i cant think
+	//       of a good way of doing it right now without loading in the model twice
+	if coffee.User.ID != user.ID {
+		server.WriteResponse(rw, http.StatusForbidden, nil)
+		return
+	}
+
+	if err := server.UploadFile(r, "image", fmt.Sprint(CoffeeImagePath, coffee.ID), &server.UploadProps{
+		Ext:  []string{".jpg", ".jpeg", ".png"},
+		Mime: []string{"image/png", "image/jpeg"},
+	}); err != nil {
+		server.WriteResponse(rw, http.StatusUnprocessableEntity, err)
+		return
+	}
+
+	server.WriteResponse(rw, http.StatusNoContent, nil)
 }
 
 type updateCoffeeRequest struct {
@@ -78,6 +121,8 @@ type updateCoffeeRequest struct {
 }
 
 func (c Controller) UpdateCoffee(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
 	id, err := server.PathID(r)
 	if err != nil {
 		server.WriteResponse(rw, http.StatusNotFound, nil)
@@ -98,6 +143,11 @@ func (c Controller) UpdateCoffee(rw http.ResponseWriter, r *http.Request) {
 	coffee, err := c.repo.FindCoffee(id)
 	if err != nil {
 		server.WriteResponse(rw, http.StatusNotFound, errors.New("Coffee not found"))
+		return
+	}
+
+	if coffee.User.ID != user.ID {
+		server.WriteResponse(rw, http.StatusForbidden, nil)
 		return
 	}
 
@@ -130,6 +180,8 @@ func (c Controller) UpdateCoffee(rw http.ResponseWriter, r *http.Request) {
 }
 
 func (c Controller) DeleteCoffee(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
 	id, err := server.PathID(r)
 	if err != nil {
 		server.WriteResponse(rw, http.StatusNotFound, nil)
@@ -142,8 +194,44 @@ func (c Controller) DeleteCoffee(rw http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if coffee.User.ID != user.ID {
+		server.WriteResponse(rw, http.StatusForbidden, nil)
+		return
+	}
+
 	if err := c.repo.DeleteCoffee(coffee); err != nil {
 		server.WriteResponse(rw, http.StatusInternalServerError, nil)
+		return
+	}
+
+	server.WriteResponse(rw, http.StatusNoContent, nil)
+}
+
+func (c Controller) UpdateRoasterImage(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
+	id, err := server.PathID(r)
+	if err != nil {
+		server.WriteResponse(rw, http.StatusNotFound, nil)
+		return
+	}
+
+	roaster, err := c.repo.FindRoaster(id)
+	if err != nil {
+		server.WriteResponse(rw, http.StatusNotFound, errors.New("Roaster not found"))
+		return
+	}
+
+	if roaster.User.ID != user.ID {
+		server.WriteResponse(rw, http.StatusForbidden, nil)
+		return
+	}
+
+	if err := server.UploadFile(r, "image", fmt.Sprint(RoasterImagePath, roaster.ID), &server.UploadProps{
+		Ext:  []string{".jpg", ".jpeg", ".png"},
+		Mime: []string{"image/png", "image/jpeg"},
+	}); err != nil {
+		server.WriteResponse(rw, http.StatusUnprocessableEntity, err)
 		return
 	}
 
@@ -155,6 +243,8 @@ type createRoasterRequest struct {
 }
 
 func (c Controller) CreateRoaster(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
 	var req createRoasterRequest
 	if err := server.UnmarshalBody(r, &req); err != nil {
 		server.WriteResponse(rw, http.StatusBadRequest, nil)
@@ -168,6 +258,7 @@ func (c Controller) CreateRoaster(rw http.ResponseWriter, r *http.Request) {
 
 	roaster := Roaster{
 		Name: req.Name,
+		User: *user,
 	}
 
 	if err := c.repo.SaveRoaster(&roaster); err != nil {
@@ -183,6 +274,8 @@ type updateRoasterRequest struct {
 }
 
 func (c Controller) UpdateRoaster(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
 	id, err := server.PathID(r)
 	if err != nil {
 		server.WriteResponse(rw, http.StatusNotFound, nil)
@@ -206,6 +299,11 @@ func (c Controller) UpdateRoaster(rw http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if roaster.User.ID != user.ID {
+		server.WriteResponse(rw, http.StatusForbidden, nil)
+		return
+	}
+
 	roaster.Name = req.Name
 
 	if err := c.repo.SaveRoaster(roaster); err != nil {
@@ -217,6 +315,8 @@ func (c Controller) UpdateRoaster(rw http.ResponseWriter, r *http.Request) {
 }
 
 func (c Controller) DeleteRoaster(rw http.ResponseWriter, r *http.Request) {
+	user := r.Context().Value("user").(*auth.User)
+
 	id, err := server.PathID(r)
 	if err != nil {
 		server.WriteResponse(rw, http.StatusNotFound, nil)
@@ -226,6 +326,11 @@ func (c Controller) DeleteRoaster(rw http.ResponseWriter, r *http.Request) {
 	roaster, err := c.repo.FindRoaster(id)
 	if err != nil {
 		server.WriteResponse(rw, http.StatusNotFound, errors.New("Roaster not found"))
+		return
+	}
+
+	if roaster.User.ID != user.ID {
+		server.WriteResponse(rw, http.StatusForbidden, nil)
 		return
 	}
 
@@ -276,40 +381,6 @@ func (c Controller) CreateFlavourProfile(rw http.ResponseWriter, r *http.Request
 
 type updateFlavourProfileRequest struct {
 	Name string `json:"name" validate:"required"`
-}
-
-func (c Controller) UpdateFlavourProfile(rw http.ResponseWriter, r *http.Request) {
-	id, err := server.PathID(r)
-	if err != nil {
-		server.WriteResponse(rw, http.StatusNotFound, nil)
-		return
-	}
-
-	var req updateFlavourProfileRequest
-	if err := server.UnmarshalBody(r, &req); err != nil {
-		server.WriteResponse(rw, http.StatusBadRequest, nil)
-		return
-	}
-
-	if err := server.ValidateRequest(req); err != nil {
-		server.WriteResponse(rw, http.StatusUnprocessableEntity, err)
-		return
-	}
-
-	flavour, err := c.repo.FindFlavourProfile(id)
-	if err != nil {
-		server.WriteResponse(rw, http.StatusNotFound, errors.New("Flavour Profile not found"))
-		return
-	}
-
-	flavour.Name = req.Name
-
-	if err := c.repo.SaveFlavourProfile(flavour); err != nil {
-		server.WriteResponse(rw, http.StatusInternalServerError, nil)
-		return
-	}
-
-	server.WriteResponse(rw, http.StatusNoContent, nil)
 }
 
 func (c Controller) DeleteFlavourProfile(rw http.ResponseWriter, r *http.Request) {
